backend/requirements.txt
Django>=3.2
djangorestframework
pandas
reportlab
matplotlib
Pillow
python-dotenv
django-cors-headers
djangorestframework-simplejwt
